ai/gptbitcoin/requirements.txt
python-dotenv
openai
pyupbit

